+++ app/lib/screens/ErrorScreen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:package_info/package_info.dart';
-import 'package:threebotlogin/main.dart';
+import 'package:threebotlogin/helpers/Globals.dart';
 
 class ErrorScreen extends StatefulWidget {
   final Widget errorScreen;
