+++ app/lib/screens/HomeScreen.dart
@@ -8,8 +8,8 @@ import 'package:threebotlogin/Events/Events.dart';
 import 'package:threebotlogin/Events/GoHomeEvent.dart';
 import 'package:threebotlogin/Events/NewLoginEvent.dart';
 import 'package:threebotlogin/Events/UniLinkEvent.dart';
+import 'package:threebotlogin/helpers/Globals.dart';
 import 'package:threebotlogin/helpers/HexColor.dart';
-import 'package:threebotlogin/main.dart';
 import 'package:threebotlogin/services/3botService.dart';
 import 'package:threebotlogin/services/socketService.dart';
 import 'package:threebotlogin/services/uniLinkService.dart';
