+++ app/lib/screens/PreferenceScreen.dart
@@ -3,12 +3,11 @@ import 'package:flutter/services.dart';
 import 'package:package_info/package_info.dart';
 import 'package:threebotlogin/Events/CloseSocketEvent.dart';
 import 'package:threebotlogin/Events/Events.dart';
+import 'package:threebotlogin/helpers/Globals.dart';
 import 'package:threebotlogin/helpers/HexColor.dart';
-import 'package:threebotlogin/main.dart';
 import 'package:threebotlogin/screens/ChangePinScreen.dart';
 import 'package:threebotlogin/services/fingerprintService.dart';
 import 'package:threebotlogin/services/openKYCService.dart';
-import 'package:threebotlogin/services/socketService.dart';
 import 'package:threebotlogin/services/userService.dart';
 import 'package:threebotlogin/widgets/CustomDialog.dart';
 import 'package:threebotlogin/widgets/EmailVerificationNeeded.dart';
